+++ config/colordescriptor.php
@@ -1,5 +1,0 @@
-<?php
-
-return [
-    'colorCount' => 6,
-];