+++ tests/Elasticsearch/Repositories/ItemRepositoryTest.php
@@ -56,4 +56,32 @@ class ItemRepositoryTest extends TestCase
             ->first();
         $this->assertTrue($items[1]->is($mostSimilar));
     }
+
+    public function testSimilarByColor()
+    {
+        /** @var Item[] $items */
+        $item = factory(Item::class)->make([
+            'colors' => ['#ff0000' => 1]
+        ]);
+
+        $similar_item = factory(Item::class)->make([
+            'colors' => ['#fe0000' => 1]
+        ]);
+
+        $dissimilar_item = factory(Item::class)->make([
+            'colors' => ['#0000ff' => 1]
+        ]);
+
+        foreach ([$item, $dissimilar_item, $similar_item] as $index_item) {
+            $this->repository->index($index_item);
+        }
+        $this->repository->refreshIndex();
+
+        $similar_by_color = $this->repository
+            ->getSimilarByColor(2, $item)
+            ->getCollection();
+
+        $this->assertTrue($similar_by_color->first()->is($similar_item));
+        $this->assertFalse($similar_by_color->contains($dissimilar_item));
+    }
 }